+++ authority-portal-backend/authority-portal-api/src/main/java/de/sovity/authorityportal/api/model/UserAuthenticationStatusDto.java
@@ -3,7 +3,7 @@ package de.sovity.authorityportal.api.model;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(enumAsRef = true)
-public enum UserOnboardingTypeDto {
-    INVITATION,
-    SELF_REGISTRATION
+public enum UserAuthenticationStatusDto {
+    AUTHENTICATED,
+    UNAUTHENTICATED
 }