+++ authority-portal-backend/authority-portal-api/src/main/java/de/sovity/authorityportal/api/model/UserRoleDto.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "Available user roles.", enumAsRef = true)
 public enum UserRoleDto {
+    UNAUTHENTICATED,
     AUTHORITY_ADMIN,
     OPERATOR_ADMIN,
     SERVICE_PARTNER_ADMIN,