+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/auth/LoggedInUser.kt
@@ -1,6 +1,7 @@
 package de.sovity.authorityportal.web.auth
 
 data class LoggedInUser(
+    val authenticated: Boolean,
     val userId: String,
     val organizationMdsId: String?,
     var roles: Set<String>