+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/pages/usermanagement/UserInvitationApiService.kt
@@ -33,7 +33,6 @@ class UserInvitationApiService {
 
         userService.createUser(
             userId = userId,
-            registrationStatus = UserRegistrationStatus.INVITED,
             mdsId = mdsId,
             onboardingType = UserOnboardingType.INVITATION,
             invitedBy = adminUserId