+++ authority-portal-backend/authority-portal-quarkus/src/test/kotlin/de/sovity/authorityportal/web/pages/userregistration/UserRegistrationStatusMapperTest.kt
@@ -9,14 +9,18 @@ class UserRegistrationStatusMapperTest {
 
     @Test
     fun testUserRegistrationStatusMapper() {
-        assertThat(UserRegistrationStatus.CREATED.toDto()).isEqualTo(UserRegistrationStatusDto.CREATED)
+        assertThat(UserRegistrationStatus.INVITED.toDto()).isEqualTo(UserRegistrationStatusDto.INVITED)
+        assertThat(UserRegistrationStatus.ONBOARDING.toDto()).isEqualTo(UserRegistrationStatusDto.ONBOARDING)
         assertThat(UserRegistrationStatus.PENDING.toDto()).isEqualTo(UserRegistrationStatusDto.PENDING)
         assertThat(UserRegistrationStatus.ACTIVE.toDto()).isEqualTo(UserRegistrationStatusDto.ACTIVE)
         assertThat(UserRegistrationStatus.REJECTED.toDto()).isEqualTo(UserRegistrationStatusDto.REJECTED)
+        assertThat(UserRegistrationStatus.DEACTIVATED.toDto()).isEqualTo(UserRegistrationStatusDto.DEACTIVATED)
 
-        assertThat(UserRegistrationStatusDto.CREATED.toDb()).isEqualTo(UserRegistrationStatus.CREATED)
+        assertThat(UserRegistrationStatusDto.INVITED.toDb()).isEqualTo(UserRegistrationStatus.INVITED)
+        assertThat(UserRegistrationStatusDto.ONBOARDING.toDb()).isEqualTo(UserRegistrationStatus.ONBOARDING)
         assertThat(UserRegistrationStatusDto.PENDING.toDb()).isEqualTo(UserRegistrationStatus.PENDING)
         assertThat(UserRegistrationStatusDto.ACTIVE.toDb()).isEqualTo(UserRegistrationStatus.ACTIVE)
         assertThat(UserRegistrationStatusDto.REJECTED.toDb()).isEqualTo(UserRegistrationStatus.REJECTED)
+        assertThat(UserRegistrationStatusDto.DEACTIVATED.toDb()).isEqualTo(UserRegistrationStatus.DEACTIVATED)
     }
 }