+++ CHANGELOG.md
@@ -49,6 +49,7 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 - Implemented CaaS integration
 - Enforced limits for free CaaS requests
 - Added accept TOS checkbox to registration form
+- New "unauthenticated page", registration now happens in the Authority Portal without Keycloak.
 
 #### Patch
 
@@ -85,7 +86,8 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
   - Added environment variable `AUTHORITY_PORTAL_FRONTEND_DSGVO_URL`
   - Added environment variable `AUTHORITY_PORTAL_FRONTEND_AVV_URL`
 
-_For further information check out our [Productive Deployment Guide](README.md#productive-deployment-guide)._
+- TODO: Create Deployment Migration Notes for the new Self-Registration flow
+
 
 #### Compatible Versions